Prepare for _Client integration testing; Requires dependancy refactor
