+++ client/Cargo.toml
@@ -26,6 +26,9 @@ tokio = { version = "1", features = ["net", "sync", "macros", "time", "rt"] }
 tokio-stream = "0.1.8"
 url = "2.1"
 
+# Necessary for testing purposes
+ya-relay-server = { path="../server"}
+
 [dev-dependencies]
 actix-rt = "2.7"
 bytesize = "1.1"
@@ -42,8 +45,6 @@ serial_test = { git = "https://github.com/tworec/serial_test.git", branch = "act
 simple-logging = "2.0"
 structopt = "0.3"
 tokio = { version = "1", features = ["fs", "io-util", "signal"] }
-
-ya-relay-server = { path="../server"}
 
 [features]
 packet-trace-enable = ["ya-packet-trace/enable"]
