+++ client/src/lib.rs
@@ -1,8 +1,6 @@
 pub mod client;
-mod dispatch;
 mod expire;
 mod registry;
-mod session;
 mod session_guard;
 mod session_manager;
 mod session_start;
@@ -16,12 +14,12 @@ pub use ya_relay_proto::*;
 pub use ya_relay_stack::*;
 
 // TODO: Exposed for ya-relay-server. Should be made private after we merge implementations.
-pub use dispatch::{dispatch, Dispatcher, Handler};
-pub use session::{Session, SessionDesc};
-pub use ya_relay_core::session::TransportType;
+pub use ya_relay_core::dispatch::{dispatch, Dispatcher, Handler};
+pub use ya_relay_core::server_session::TransportType;
+pub use ya_relay_core::session::{Session, SessionDesc};
 
 // Experimental
-mod _client;
+pub mod _client;
 mod _config;
 mod _direct_session;
 mod _dispatch;
