+++ client/src/testing.rs
@@ -1,6 +1,5 @@
 pub mod accessors;
 pub mod forwarding_utils;
-#[cfg(test)]
 pub mod init;
 
-pub use crate::session::Session;
+pub use ya_relay_core::session::Session;
