+++ crates/core/src/lib.rs
@@ -1,8 +1,10 @@
 pub mod challenge;
 pub mod crypto;
+pub mod dispatch;
 pub mod error;
 pub mod identity;
 pub mod key;
+pub mod server_session;
 pub mod session;
 pub mod sync;
 pub mod udp_stream;
